Module Sc_values

An interface for representing C values of any type.

module TYPES : sig ... end

include module type of TYPES with type encoding_params = TYPES.encoding_params and type struct_field = TYPES.struct_field and type field_access = TYPES.field_access and type memory_validity = TYPES.memory_validity and type ptr_attributes = TYPES.ptr_attributes and type 'a ctyp_consumer = 'a TYPES.ctyp_consumer and type 'a ptr_attributes_folder = 'a TYPES.ptr_attributes_folder and type 'a ptr_access_path_folder = 'a TYPES.ptr_access_path_folder and type 'a cptr_consumer = 'a TYPES.cptr_consumer

type cside

Abstract type for any C-side value

type _ ctyp

Abstract representation of C types

type struct_

Types of some C types

type union_

type array_

type 'a ctyp_consumer = 'a TYPES.ctyp_consumer = {

f : 'k. 'k Ctypes.typ -> 'a;

}

Polymorphic function over internal C types

type typdecls

Type of typing declarations for the naming and definition of composed types.

type encoding_params = TYPES.encoding_params = {

max_ptr_array_length : int;
(*
Maximal size of C arrays directly reachable via pointers.
*)

max_non_nil_ptr_depth : int;
(*
Maximal depth at which pointers may be non-NULL, measured in number of pointer dereferences.
*)

max_cstring_length : int;
(*
Maximal size of C strings, excluding any additional terminating '\000'.
*)

}

type ptr_attributes = TYPES.ptr_attributes = {

pv : memory_validity;

depth : int;

}

View of pointer attributes, exposed during structure traversals below

and memory_validity = TYPES.memory_validity

and field_access = TYPES.field_access = {

ap_suffix : Sc_C.Types.access_path_suffix;

length_field : struct_field;

}

and struct_field = TYPES.struct_field

type 'a ptr_attributes_folder = 'a TYPES.ptr_attributes_folder

type 'a access_path_folder = (Sc_C.Types.access_path_suffix option -> 'a -> 'a) ctyp_consumer

type 'a ptr_access_path_folder = 'a TYPES.ptr_access_path_folder

type 'a field_folder = (string -> 'a -> 'a) ctyp_consumer

type 'a ptr_field_folder = (string -> memory_validity -> 'a -> 'a) ctyp_consumer

type 'a visit_action =

| Do_children of 'a

| Do_children_and_then of 'a * 'a -> 'a

| Skip_children of 'a

type 'a access_path_visitor = (Sc_C.Types.access_path_suffix option -> 'a -> 'a visit_action) ctyp_consumer

type 'a ptr_access_path_visitor = (Sc_C.Types.access_path_suffix option -> memory_validity -> 'a -> 'a visit_action) ctyp_consumer

type 'a cptr_consumer = 'a TYPES.cptr_consumer = {

p : 'k. 'k Ctypes.typ -> 'k Ctypes.ptr -> ?len:int -> ptr_attributes -> 'a;

}

type specification_error =

| Unknown_field of {

field_name : string;

aggreg_type_tag : string;

}

| Invalid_length_field of {

field_name : string;

field_typ : Ctypes_static.boxed_typ;

}

| Unsupported_type of Cil.typ

| Unknown_enum of Cil.enuminfo

| Unknown_comp of Cil.compinfo

exception SPECIFICATION_ERROR of specification_error

type literal_binding =

| LBStr of string

| LBMap of literal_binding Basics.StrMap.t

| LBArr of literal_binding array

| LBLst of literal_binding list

| LBNil
(*
NULL reference
*)

| LBRef of literal_binding
(*
reference to bindings
*)

Type of bindings to C values directly from C string literals; given to VAL.assign_from_literal below.

type literal_binding_error =

| Unexpected_literal of {

binding : literal_binding;

assigned : (string * Ctypes_static.boxed_typ) option;

msg : string;

}

| Unsupported_assignment of assignment_info

| Unknown_enum of {

enumname : string;

assignment : assignment_info;

}

| Incompatible_array_length of {

expected_len : int;

rvalue : literal_binding array;

}

| Bad_array_literal_length of {

max_len : int;

given_len : int;

assignment : assignment_info;

}

| Bad_cstring_literal_length of {

string : string;

params : encoding_params;

assignment : assignment_info;

}

and assignment_info = {

binding : literal_binding;

assigned : string * Ctypes_static.boxed_typ;

}

exception LITERAL_BINDING_ERROR of literal_binding_error

Exception raised in assignments with literal bindings

val empty_typdecls : typdecls

An immutable, empty set of type declarations.

val append_global_cil_declaration_to_typdecls : Cil.global -> typdecls -> typdecls

Note: ignores anything other than type declarations

val append_global_cil_declarations_to_typdecls : Cil.global list -> typdecls -> typdecls

val typdecls_from_cil_file : Cil.file -> typdecls

typdecls_from_cil_file cil returns a set of typing declarations built from all global type definitions in cil.

val sizeof_cil_typ : ?typdecls:typdecls -> Cil.typ -> int

sizeof_cil_typ is the size of the given Cil. type, in bytes (equivalent to [sizeof(t)]), possibly given some type declarations.

val struct_field_typ : struct_field -> Ctypes_static.boxed_typ

val pp_literal_binding : literal_binding Fmt.t

Prints raw literal bindings (for debug)

val pp_field_access : TYPES.field_access Fmt.t

val pp_literal_binding_error : TYPES.literal_binding_error Fmt.t

val pp_specification_error : TYPES.specification_error Fmt.t

module type ENCODING_PARAMS = sig ... end

Parameters for binary encoding

module type ENCODING = sig ... end

High-level encoding specification. This only gathers parameters as no further details about the encoding should be exported.

module type VALUE_PRINTER = sig ... end

The type of represented values, parameterized by encoding parameters.

module Printer : sig ... end

Helpers to print C types and variable declarations

module Encoding (_ : sig ... end) : sig ... end

Encoding-specific low-level routines

module type VAL = sig ... end

Encapsulated typed values

module type REPR = sig ... end

Representation of typed values that is parametric in encoding parameters

module type AGGREG_TYP = sig ... end

Representation of C composed types (structure, union, arrays). These can be used to interpret any sufficiently long OCaml character string as a memory representation of a value for such a type.

module type NAMED_MAPPING = sig ... end

Representation of C structured types (structure or union). These can be built directly from a list of CIL field descriptions.

module Struct : sig ... end

Declaration and representation of C struct types.

module Union : sig ... end

Declaration and representation of C unions.

module BoxedArray : sig ... end

Declaration and representation of C arrays.

Other utilities for easy access to Ctypes

val has_pointers : 'k Ctypes.typ -> bool

has_pointers t is true iff the C type t is composed of at least one pointer field.

module Raw_decoder : sig ... end