Module Val.Printer

type literal_memory = literal_assignments * literal_heap

A literal memory consists of initialized typed variables, along with a literal heap

and literal_assignments

and literal_heap

and initialization_block

type c_code = {

pp_heap : static:bool -> Basics.PPrt.pu;
(*
pp_heap ~static ppf uses ppf to print parseable C code that declares and initializes heap memory cells. Initialized declarations (i.e, `typ cell = lit;') are used if static = true; calls to `memcpy` are used otherwise (i.e, `typ cell; (void) memcpy (&cell, ...);`).
*)

pp_globals : Basics.PPrt.pu;
(*
prints assignment of already declared variables
*)

pp_locals : Basics.PPrt.pu;
(*
prints declarations and initializations for local variables
*)

}

C code used to declare and initialize C memory

val literal_memory_as_c_code : globals:Sc_C.Types.vars -> locals:Sc_C.Types.vars -> literal_memory -> c_code

literal_memory_as_c_code ~globals ~locals instructions constructs printers that output the given literal memory as C code. This codes does not make use of any dynamic memory. In particular, it does not use `malloc` --- it may however use `memcpy` to initialize large objects.

val initialization_block_as_c_code : globals:Sc_C.Types.vars -> locals:Sc_C.Types.vars -> initialization_block -> c_code

initialization_block_as_c_code ~globals ~locals instructions constructs printers that output instructions in C. This code makes use of `malloc` and potentially `memcpy` to allocate and initialize heap objects. Note pp_heap is a no-op (as the code of instructions does not rely on any static declaration or initialization of heap objects).