Module Sc_values.Encoding

Encoding-specific low-level routines

Parameters

module _ : sig ... end

Signature

val pp_c_ptr_decode : (string * string * string * string) Fmt.t

pp_c_ptr_decode ppf (v, ub, d) prints using ppf a C expression that computes a length of array encoded by a given unsigned long variable v, at a depth encoded by an unsigned int d. The lower/upper bound for the length is lb/ub.

val pp_c_constrained_carray_length : (string * string * string * string) Fmt.t

pp_c_constrained_carray_length ppf (v, size, ub, d) prints using ppf a C expression that computes the actual a length of an array with "given" length *v, at a depth encoded by an unsigned int d. ub is the upper bound on for the length.

The main purpose of the resulting expression is to take the depth into accound to for the length to zero when the depth exceeds max_non_nil_ptr_depth.

val pp_c_cstring_buff_length_decode : (string * string) Fmt.t

pp_cstring_buff_length_decode fmt (v, d) prints using fmt a C expression that computes a length of C character string, including the trailing '\000', encoded by a given unsigned long variable v, at a depth encoded by an unsigned int d.

val max_pointed_size : 'k Ctypes.typ -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> int

max_pointed_size typ pv is the maximum size in bytes of an array of elements of type typ that may be pointed to by a pointer with validity pv.

val ctyp_fold_access_paths : _ Ctypes_static.typ -> ?max_depth:int -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

ctyp_fold_access_paths ctyp ~max_depth ~f_prm ~f_ptr acc folds f_prm and f_ptr over every canonical access path suffix that is may be a valid memory access from a value of type ctyp. f_ptr is used for explicit pointers; f_prm for every other type of data (including plain arrays); pointer types are visited recursively (up to max_depth, which is by default params.max_non_nil_ptr_depth).

Note: use max_depth=0 to prevent traversal of pointers.

val ctyp_fold_direct_access_paths : _ Ctypes.typ -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

ctyp_fold_direct_access_paths ctyp ~f_prm ~f_ptr acc folds f_prm and f_ptrf over every field defined in type ctyp; referenced types are NOT visited recursively.

This is a shorthand for ctyp_fold_access_paths ~max_depth:0 ?f_prm ?f_ptr.

val ctyp_fold_reachable_pointers : _ Ctypes_static.typ -> ?max_depth:int -> f: (Sc_values__.Cvalues.TYPES.ptr_attributes -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> merge:('a NEL.t -> 'a) -> 'a -> 'a

ctyp_fold_reachable_pointers ctyp ~max_depth ~f ~merge acc folds f over every pointer field defined in type ctyp; pointed types are visited recursively (up to and including max_depth), and results of accumulation over union fields are computed using merge. max_depth defaults to max_non_nil_ptr_depth (given via encoding parameters).

val ctyp_visit_access_paths : _ Ctypes_static.typ -> ?max_depth:int -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a Sc_values__.Cvalues.visit_action) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a Sc_values__.Cvalues.visit_action) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

ctyp_visit_access_paths ctyp ~max_depth ~f_prm ~f_ptr acc visits every (canonicalized) access path that may be valid on a value of type ctyp; pointers are visited using f_ptr, while every other type of value is visited using f_prm.

The value returned by these visiting functions on any path ap drives the traversal: when Do_children_and_then (acc, f) is returned, any type of memory cell that may be accessed by appending a field or array indexing or pointer access suffix to ap is visited recursively (unless max_depth is reached when traversing a pointer type) with the accumulated value acc; then the resulting accumulated value is passed to f. Do_children acc is equivalent to Do_children_and_then (acc, Fun.id). When Skip_children acc is returned, the result of the traversal of ap is acc.