Module Sc_values.TYPES

type cside

Abstract type for any C-side value

type _ ctyp

Abstract representation of C types

type struct_

Types of some C types

type union_

type array_

type 'a ctyp_consumer = {

f : 'k. 'k Ctypes.typ -> 'a;

}

Polymorphic function over internal C types

type typdecls

Type of typing declarations for the naming and definition of composed types.

type encoding_params = {

max_ptr_array_length : int;
(*
Maximal size of C arrays directly reachable via pointers.
*)

max_non_nil_ptr_depth : int;
(*
Maximal depth at which pointers may be non-NULL, measured in number of pointer dereferences.
*)

max_cstring_length : int;
(*
Maximal size of C strings, excluding any additional terminating '\000'.
*)

}

type ptr_attributes = {

pv : memory_validity;

depth : int;

}

View of pointer attributes, exposed during structure traversals below

and memory_validity = [

| `Carray_with_bound_length of int

| `Carray_with_length_field of field_access

| `Cstring

]

and field_access = {

ap_suffix : Sc_C.Types.access_path_suffix;

length_field : struct_field;

}

and struct_field

type 'a ptr_attributes_folder = (ptr_attributes -> 'a -> 'a) ctyp_consumer

type 'a access_path_folder = (Sc_C.Types.access_path_suffix option -> 'a -> 'a) ctyp_consumer

type 'a ptr_access_path_folder = (Sc_C.Types.access_path_suffix option -> memory_validity -> 'a -> 'a) ctyp_consumer

type 'a field_folder = (string -> 'a -> 'a) ctyp_consumer

type 'a ptr_field_folder = (string -> memory_validity -> 'a -> 'a) ctyp_consumer

type 'a visit_action =

| Do_children of 'a

| Do_children_and_then of 'a * 'a -> 'a

| Skip_children of 'a

type 'a access_path_visitor = (Sc_C.Types.access_path_suffix option -> 'a -> 'a visit_action) ctyp_consumer

type 'a ptr_access_path_visitor = (Sc_C.Types.access_path_suffix option -> memory_validity -> 'a -> 'a visit_action) ctyp_consumer

type 'a cptr_consumer = {

p : 'k. 'k Ctypes.typ -> 'k Ctypes.ptr -> ?len:int -> ptr_attributes -> 'a;

}

type specification_error =

| Unknown_field of {

field_name : string;

aggreg_type_tag : string;

}

| Invalid_length_field of {

field_name : string;

field_typ : Ctypes_static.boxed_typ;

}

| Unsupported_type of Cil.typ

| Unknown_enum of Cil.enuminfo

| Unknown_comp of Cil.compinfo

exception SPECIFICATION_ERROR of specification_error

type literal_binding =

| LBStr of string

| LBMap of literal_binding Basics.StrMap.t

| LBArr of literal_binding array

| LBLst of literal_binding list

| LBNil
(*
NULL reference
*)

| LBRef of literal_binding
(*
reference to bindings
*)

Type of bindings to C values directly from C string literals; given to VAL.assign_from_literal below.

type literal_binding_error =

| Unexpected_literal of {

binding : literal_binding;

assigned : (string * Ctypes_static.boxed_typ) option;

msg : string;

}

| Unsupported_assignment of assignment_info

| Unknown_enum of {

enumname : string;

assignment : assignment_info;

}

| Incompatible_array_length of {

expected_len : int;

rvalue : literal_binding array;

}

| Bad_array_literal_length of {

max_len : int;

given_len : int;

assignment : assignment_info;

}

| Bad_cstring_literal_length of {

string : string;

params : encoding_params;

assignment : assignment_info;

}

and assignment_info = {

binding : literal_binding;

assigned : string * Ctypes_static.boxed_typ;

}

exception LITERAL_BINDING_ERROR of literal_binding_error

Exception raised in assignments with literal bindings