Module Sc_values.Union

Declaration and representation of C unions.

Declaration and representation of C unions.

type typ

C composed type.

type value = Sc_values__.Cvalues.cside Ctypes.union

C value built for the composed type.

val size : typ -> int

size typ returns the size in bytes of the memory representation for values of the composed type typ.

val definition : Stdlib.Format.formatter -> typ -> unit

print fmt typ prints the C definition of the composed type typ on the formatter fmt.

val print : Stdlib.Format.formatter -> typ -> unit

val declare : string -> typ -> Sc_values__.Cvalues.typdecls -> Sc_values__.Cvalues.typdecls

declare name typ typdecls declares a typ with name name into typdecls. Any previous binding of name in typdecls is erased.

val as_ctyp : typ -> f:'a Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a

as_ctyp typ f applies f on the `ctypes` representation of typ.

exception Invalid_value_representation of typ * string option

Exception raised by Val.of_string when it is given a string that is too short to hold a value of type typ.

module type REPR = sig ... end

Bound representation types

module Repr (Params : sig ... end) : REPR with module Params = Params

Parametric representation of typed values

val from_cil_fields : ?typdecls:Sc_values__.Cvalues.typdecls -> string -> Sc_C.Types.vars -> typ

from_cil_fields ~typdecls name fields defines a composed type based on a list of named CIL fields.